refactor(api): move cosmos DB util
