--- api/db/index.d.ts
@@ -1,12 +1,0 @@
-import type { NotificationSchema } from './notification'
-import type { ScoreSchema } from './scores'
-import type { CourseSchema, SongSchema } from './songs'
-import type { UserSchema } from './users'
-
-export type {
-  CourseSchema,
-  NotificationSchema,
-  ScoreSchema,
-  SongSchema,
-  UserSchema,
-}
