+++ api/db/index.ts
@@ -1,5 +1,18 @@
 import { Container, CosmosClient } from '@azure/cosmos'
 
+import type { NotificationSchema } from './notification'
+import type { ScoreSchema } from './scores'
+import type { CourseSchema, SongSchema } from './songs'
+import type { UserSchema } from './users'
+
+export type {
+  CourseSchema,
+  NotificationSchema,
+  ScoreSchema,
+  SongSchema,
+  UserSchema,
+}
+
 type ContainerName = 'Scores' | 'Songs' | 'Users' | 'Notification'
 
 /* eslint-disable node/no-process-env */
