+++ api/deleteChartScore/index.test.ts
@@ -3,8 +3,7 @@ import { mocked } from 'ts-jest/utils'
 
 import { describeIf } from '../__tests__/util'
 import { getClientPrincipal, getLoginUserInfo } from '../auth'
-import { getConnectionString, getContainer } from '../cosmos'
-import { ScoreSchema } from '../db'
+import { getConnectionString, getContainer, ScoreSchema } from '../db'
 import deleteChartScore from '.'
 
 jest.mock('../auth')
