+++ api/existsUser/index.test.ts
@@ -3,8 +3,7 @@ import { mocked } from 'ts-jest/utils'
 
 import { describeIf } from '../__tests__/util'
 import { getClientPrincipal } from '../auth'
-import { getConnectionString, getContainer } from '../cosmos'
-import type { UserSchema } from '../db'
+import { getConnectionString, getContainer, UserSchema } from '../db'
 import existsUser from '.'
 
 jest.mock('../auth')
