+++ api/getChartScore/index.ts
@@ -2,8 +2,7 @@ import type { SqlParameter } from '@azure/cosmos'
 import type { Context, HttpRequest } from '@azure/functions'
 
 import { getClientPrincipal } from '../auth'
-import { getContainer } from '../cosmos'
-import type { ScoreSchema, UserSchema } from '../db'
+import { getContainer, ScoreSchema, UserSchema } from '../db'
 import type { NotFoundResult, SuccessResult } from '../function'
 
 type Score = Omit<ScoreSchema, 'id' | 'isPublic'>
