+++ api/getCurrentUser/index.test.ts
@@ -2,8 +2,7 @@ import { mocked } from 'ts-jest/utils'
 
 import { describeIf } from '../__tests__/util'
 import { getClientPrincipal } from '../auth'
-import { getConnectionString, getContainer } from '../cosmos'
-import type { UserSchema } from '../db'
+import { getConnectionString, getContainer, UserSchema } from '../db'
 import getCurrentUser from '.'
 
 jest.mock('../auth')
