+++ api/getNotificationInfo/index.ts
@@ -1,7 +1,6 @@
 import type { Context } from '@azure/functions'
 
-import { getContainer } from '../cosmos'
-import type { NotificationSchema } from '../db'
+import { getContainer, NotificationSchema } from '../db'
 import type { NotFoundResult, SuccessResult } from '../function'
 
 /** Get notification that match the specified ID. */
