+++ api/getNotificationList/index.test.ts
@@ -1,6 +1,5 @@
 import { describeIf } from '../__tests__/util'
-import { getConnectionString, getContainer } from '../cosmos'
-import type { NotificationSchema } from '../db'
+import { getConnectionString, getContainer, NotificationSchema } from '../db'
 import getNotification from '.'
 
 describe('GET /api/v1/notification', () => {
