+++ api/getUserList/index.ts
@@ -2,7 +2,7 @@ import { SqlParameter } from '@azure/cosmos'
 import type { HttpRequest } from '@azure/functions'
 
 import { getClientPrincipal } from '../auth'
-import { getContainer } from '../cosmos'
+import { getContainer } from '../db'
 import { areaCodeList } from '../db/users'
 import type { SuccessResult } from '../function'
 import type { User } from '../user'
