+++ api/postUserInfo/index.ts
@@ -1,7 +1,7 @@
 import type { HttpRequest } from '@azure/functions'
 
 import { getClientPrincipal } from '../auth'
-import { getContainer } from '../cosmos'
+import { getContainer } from '../db'
 import { areaCodeList, UserSchema } from '../db/users'
 import type {
   BadRequestResult,
